+++ integrations/llama/run.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-
-CURRENT_DIR=/home/jon/git/llm_server/integrations/llama
-cd /home/jon/git/llm_server/integrations/llama
-
-/home/jon/git/llm_server/integrations/llama/models/./Qwen3-Embedding-4B-Q8_0.sh
-
-